Break out toEqualJSON
