+++ pg-toolbox/__test__/utils.ts
@@ -20,3 +20,40 @@ export const setupTestDatabase = (): Database => {
 
   return db
 }
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace jest {
+    interface Matchers<R> {
+      toEqualJSON(v: unknown): R
+    }
+  }
+}
+
+export const extendExpect = () => {
+  expect.extend({
+    // Expect the two values to equal when JSON-stringified.
+    toEqualJSON(received: unknown, expected: unknown) {
+      const pass = this.equals(
+        JSON.stringify(received),
+        JSON.stringify(expected),
+      )
+
+      const message = () => {
+        const expectedPrefix = pass ? 'not ' : ''
+        return [
+          this.utils.matcherHint('toEqualJSON', undefined, undefined, {
+            comment: 'equality after JSON.stringify',
+            isNot: this.isNot,
+            promise: this.promise,
+          }),
+          '',
+          'Expected: ' + expectedPrefix + this.utils.printExpected(expected),
+          'Received: ' + this.utils.printReceived(received),
+        ].join('\n')
+      }
+
+      return { actual: received, message, pass }
+    },
+  })
+}
